fix(pu): Harden UniZero Atari segment configs and replay lifecycle
- Split the concise stable segment baseline from the experimental launcher while preserving async launch helpers
- Add max-env-step and collect-temperature overrides with validated reanalysis compatibility defaults
- Preserve per-env in-flight segments across collection boundaries and track real transitions before bootstrap padding
- Align replay priorities, sampling indices, and retention with real transition counts for PER-on GSL200 runs
- Add focused config, resume, collector lifecycle, and game-buffer regression coverage
